jirs/requirements.txt
python-dotenv
openai
jiraone
from_root
-e .